3.2 Added pagination for voluminous queries.
