+++ EventHub-IntexSoft/src/main/java/com/Eventhub/EventHubIntexSoft/config/ApplicationConfiguration.java
@@ -1,8 +1,7 @@
-package com.Eventhub.EventHubIntexSoft.security;
-
-import java.util.Optional;
+package com.Eventhub.EventHubIntexSoft.config;
 
 import com.Eventhub.EventHubIntexSoft.service.UserService;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -17,7 +16,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
 @RequiredArgsConstructor
-public class ApplicationConfig {
+public class ApplicationConfiguration {
   private final UserService userService;
 
   @Bean
