+++ EventHub-IntexSoft/src/main/java/com/Eventhub/EventHubIntexSoft/controller/UserController.java
@@ -1,21 +1,17 @@
 package com.Eventhub.EventHubIntexSoft.controller;
 
-import com.Eventhub.EventHubIntexSoft.dto.AuthRequestDto;
-import com.Eventhub.EventHubIntexSoft.dto.JwtResponseDto;
-import com.Eventhub.EventHubIntexSoft.dto.RefreshTokenRequestDto;
 import com.Eventhub.EventHubIntexSoft.dto.UserDto;
-import com.Eventhub.EventHubIntexSoft.entity.RefreshToken;
 import com.Eventhub.EventHubIntexSoft.exception.EmptyDtoFieldException;
 import com.Eventhub.EventHubIntexSoft.exception.FormatException;
 import com.Eventhub.EventHubIntexSoft.exception.NonUniqValueException;
 import com.Eventhub.EventHubIntexSoft.exception.NotFoundException;
-import com.Eventhub.EventHubIntexSoft.service.AuthenticationService;
-import com.Eventhub.EventHubIntexSoft.service.JwtService;
-import com.Eventhub.EventHubIntexSoft.service.RefreshTokenService;
 import com.Eventhub.EventHubIntexSoft.service.UserService;
 import com.Eventhub.EventHubIntexSoft.validator.UserValidator;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -28,8 +24,10 @@ public class UserController {
   private final UserValidator userValidator;
 
   @GetMapping("/all")
-  public ResponseEntity<List<UserDto>> allUsers() {
-    return ResponseEntity.ok(userService.getAllUsers());
+  public ResponseEntity<Page<UserDto>> allUsers(
+      @RequestParam(value = "offset", defaultValue = "0") @Min(0) Integer offset,
+      @RequestParam(value = "limit", defaultValue = "5") @Min(1) @Max(100) Integer limit) {
+    return ResponseEntity.ok(userService.getAllUsers(offset, limit));
   }
 
   @PostMapping
