+++ EventHub-IntexSoft/src/main/java/com/Eventhub/EventHubIntexSoft/service/EventService.java
@@ -2,10 +2,12 @@ package com.Eventhub.EventHubIntexSoft.service;
 
 import com.Eventhub.EventHubIntexSoft.dto.EventDto;
 import com.Eventhub.EventHubIntexSoft.entity.Event;
+import org.springframework.data.domain.Page;
+
 import java.util.List;
 
 public interface EventService {
-  List<EventDto> getAllEvents();
+  Page<EventDto> getAllEvents(Integer offset, Integer limit);
 
   EventDto createEvent(EventDto eventDto);
