+++ EventHub-IntexSoft/src/main/java/com/Eventhub/EventHubIntexSoft/service/Impl/CommentServiceImpl.java
@@ -7,12 +7,13 @@ import com.Eventhub.EventHubIntexSoft.repository.CommentRepository;
 import com.Eventhub.EventHubIntexSoft.service.CommentService;
 import java.beans.FeatureDescriptor;
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.BeanWrapper;
 import org.springframework.beans.BeanWrapperImpl;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,8 +24,10 @@ public class CommentServiceImpl implements CommentService {
   private final CommentRepository commentRepository;
   private final CommentMapper commentMapper;
 
-  public List<CommentDto> getAllComments() {
-    return commentMapper.toDtoList(commentRepository.findAll());
+  public Page<CommentDto> getAllComments(Integer offset, Integer limit) {
+    return commentRepository
+        .findAll(PageRequest.of(offset, limit))
+        .map(commentMapper::toCommentDto);
   }
 
   public CommentDto createComment(CommentDto commentDto) {
