+++ EventHub-IntexSoft/src/main/java/com/Eventhub/EventHubIntexSoft/service/ParticipantService.java
@@ -1,10 +1,12 @@
 package com.Eventhub.EventHubIntexSoft.service;
 
 import com.Eventhub.EventHubIntexSoft.dto.ParticipantDto;
+import org.springframework.data.domain.Page;
+
 import java.util.List;
 
 public interface ParticipantService {
-  List<ParticipantDto> getAllParticipants();
+  Page<ParticipantDto> getAllParticipants(Integer offset, Integer limit);
 
   ParticipantDto createParticipant(ParticipantDto participantDto);
